feat: add run command to Makefile and introduce Recommendation types
