+++ src/strategy/__init__.py
@@ -2,10 +2,6 @@
 Strategy module for resource optimization.
 """
 
-from dataclasses import dataclass
-from enum import Enum
-from typing import List
-
 from .adaptive_strategy import AdaptiveStrategy as AdaptiveStrategy
 from .base_strategy import BaseStrategy as BaseStrategy
 from .basic_strategy import BasicStrategy as BasicStrategy
@@ -19,39 +15,10 @@ from .quantile_regression_strategy import (
 from .strategy_factory import StrategyFactory as StrategyFactory
 from .time_aware_strategy import TimeAwareStrategy as TimeAwareStrategy
 from .trend_aware_strategy import TrendAwareStrategy as TrendAwareStrategy
+from .types import RecommendationConfig as RecommendationConfig
+from .types import RecommendationStrategy as RecommendationStrategy
 from .workload_aware_strategy import WorkloadAwareStrategy as WorkloadAwareStrategy
 
-
-class RecommendationStrategy(Enum):
-    BASIC = "basic"
-    PERCENTILE = "percentile"
-    TIME_AWARE = "time_aware"
-    TREND_AWARE = "trend_aware"
-    WORKLOAD_AWARE = "workload_aware"
-    ADAPTIVE = "adaptive"
-    QUANTILE_REGRESSION = "quantile_regression"
-    MOVING_AVERAGE = "moving_average"
-    PMDARIMA = "pmdarima"
-    PROPHET = "prophet"
-    ENSEMBLE = "ensemble"
-
-
-@dataclass
-class RecommendationConfig:
-    strategy: RecommendationStrategy
-    cpu_percentile: float = 95.0
-    memory_buffer: float = 1.15
-    min_cpu_cores: float = 0.01
-    min_memory_bytes: float = 100 * 1024 * 1024  # 100Mi
-    business_hours_start: int = 9
-    business_hours_end: int = 17
-    business_days: List[int] = (0, 1, 2, 3, 4)  # Monday = 0
-    trend_threshold: float = 0.1
-    high_variance_threshold: float = 0.5
-    history_window_hours: int = 24  # Default to 24 hours of historical data
-
-
-# Import submodules after types are defined to avoid circular imports
 __all__ = [
     "BaseStrategy",
     "BasicStrategy",
@@ -65,4 +32,6 @@ __all__ = [
     "ProphetStrategy",
     "EnsembleStrategy",
     "StrategyFactory",
+    "RecommendationConfig",
+    "RecommendationStrategy",
 ]
